Project-16.1-MBTA/flask_app/requirements.txt
Flask==3.1.2
requests==2.32.5
mysql-connector-python==9.5.0
